api/requirements.txt
fastapi==0.68.0
python-multipart==0.0.5
pillow==9.5.0
torch==2.0.0
ultralytics==8.0.0
uvicorn==0.15.0
python-dotenv==0.19.0